Updated Client to Make a Gets Largest Server Method
The client code has been updated so that the code for the getting the server information has been moved to its own function that is called now before the main while loop. An additional function has been created which satisfies the algorithm allToLargest where the server list info arrayList is looped and split and held in a temporary array which updates through each loop of the list. New ServerInfo objects are created with each loop that have their attributes assigned to the corresponding positions of the array. These new objects are then added into their own arrayList where it is looped through and the corecount is checked to see which is the largest the global int are then updated to hold largest corecount, server ID and server type.